OneDrive/Desktop/Apna development/clinic-dash-demo-main/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pydantic
python-multipart
reportlab
python-dotenv
email-validator
